+++ src/plone/restapi/deserializer/dxcontent.py
@@ -62,6 +62,21 @@ class DeserializeFromJson(OrderingMixin, object):
                     if not self.check_permission(write_permissions.get(name)):
                         continue
 
+                    # set the field to missing_value if we receive null
+                    if data[name] is None:
+                        if not field.required:
+                            dm.set(field.missing_value)
+                        else:
+                            errors.append({
+                                'field': field.__name__,
+                                'message': (
+                                    '{} is a required field.'.format(
+                                        field.__name__
+                                    ),
+                                    'Setting it to null is not allowed.'
+                                )})
+                        continue
+
                     # Deserialize to field value
                     deserializer = queryMultiAdapter(
                         (field, self.context, self.request),