+++ src/plone/restapi/tests/test_content_patch.py
@@ -27,7 +27,8 @@ class TestContentPatch(unittest.TestCase):
         self.portal.invokeFactory(
             'Document',
             id='doc1',
-            title='My Document'
+            title='My Document',
+            description='Some Description'
         )
         wftool = getToolByName(self.portal, 'portal_workflow')
         wftool.doActionFor(self.portal.doc1, 'publish')
@@ -43,6 +44,41 @@ class TestContentPatch(unittest.TestCase):
         self.assertEqual(204, response.status_code)
         transaction.begin()
         self.assertEqual("Patched Document", self.portal.doc1.Title())
+
+    def test_patch_document_will_delete_value_with_null(self):
+        self.assertEqual(self.portal.doc1.description, 'Some Description')
+        response = requests.patch(
+            self.portal.doc1.absolute_url(),
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+            data='{"description": null}',
+        )
+        transaction.commit()
+
+        # null will set field.missing_value which is u'' for the field
+        self.assertEqual(204, response.status_code)
+        self.assertEqual(u'', self.portal.doc1.description)
+
+    def test_patch_document_will_not_delete_value_with_null_if_required(self):
+        response = requests.patch(
+            self.portal.doc1.absolute_url(),
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+            data='{"title": null}',
+        )
+        transaction.commit()
+
+        # null will set field.missing_value which is u'' for the field
+        self.assertEqual(400, response.status_code)
+        self.assertTrue("\'field\': \'title\'" in response.text)
+        self.assertTrue(
+            'title is a required field.'
+            in response.text
+        )
+        self.assertTrue(
+             'Setting it to null is not allowed.'
+             in response.text
+        )
 
     def test_patch_document_with_representation(self):
         response = requests.patch(